pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oju"
version = "0.1.0"
description = "A multi-agent framework supporting multiple LLM providers with structured prompts"
readme = "README.md"
authors = [
    {name = "Ojas Aklecha", email = "ojasaklechayt@gmail.com"},
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["llm", "ai", "agent", "framework", "openai", "anthropic", "gemini"]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/ojasaklechayt/oju"
"Bug Tracker" = "https://github.com/ojasaklechayt/oju/issues"
Source = "https://github.com/ojasaklechayt/oju"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.5b2",
    "isort>=5.0.0",
    "mypy>=0.900",
    "pylint>=2.8.0",
]

[project.scripts]
oju = "oju.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["oju*"]

[tool.setuptools.package-data]
oju = ["prompts/**/*.txt"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = false
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
show_error_context = true
pretty = true
show_column_numbers = true
show_error_end = true